backend/requirements.txt
python-multipart==0.0.9
pytz==2024.1
authing==4.5.18
websockets==12.0
opentelemetry-api==1.25.0
motor==3.4.0
prometheus_client==0.20.0
Jinja2==3.1.4
oss2==2.18.4
APScheduler>=3.11.0
PyJWT==2.8.0
minio
starlette==0.46.2
nest_asyncio==1.6.0
uvicorn==0.34.2
fastapi==0.115.12
apscheduler==3.11.0
PyYAML==6.0.2
python-json-logger==2.0.7
aiohttp==3.9.3
protobuf==4.25.3
websocket-client==1.7.0